Updated data_processing. Less options on run
Changed clustering method to have it make more sense and be more efficient. There is the option to get vectors with the 'argmax' of clusters if you would like. language_vectors method returns vectors which can be used for comparison. It should be super easy to finish now. There are less options on the run.py to correspond. use methods independently from run.py